% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_adstock.R
\name{geometric_adstock}
\alias{geometric_adstock}
\title{Calculates geometric adstock}
\usage{
geometric_adstock(x, lambda, log = FALSE, fischer = FALSE, log_first = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{lambda}{The carryover parameter to consider when calculating ad stock.}

\item{log}{If TRUE result will be returned as ln(. + 1).}

\item{fischer}{If TRUE initital ad stock will be estimated according to Fischer (XXX).}

\item{log_first}{If TRUE log will be calculated based on the flow variables.}
}
\value{
The ad stock transformation of x
}
\description{
Calculates geometric adstock
}
\examples{
adstock_x <- geometric_adstock(x, lambda = 0.8, log = TRUE, fischer = FALSE, log_first = FALSE)
}
